# Conflicts:
#	src/BackupDestination/BackupCollection.php
#	src/BackupServiceProvider.php
#	src/Commands/CleanupCommand.php
#	src/Commands/ListCommand.php
#	src/Commands/MonitorCommand.php
#	src/Notifications/BaseNotification.php
#	src/Notifications/EventHandler.php
#	src/Notifications/Notifications/BackupHasFailed.php
#	src/Notifications/Notifications/BackupWasSuccessful.php
#	src/Notifications/Notifications/CleanupHasFailed.php
#	src/Notifications/Notifications/CleanupWasSuccessful.php
#	src/Notifications/Notifications/HealthyBackupWasFound.php
#	src/Notifications/Notifications/UnhealthyBackupWasFound.php
#	src/Tasks/Backup/BackupJob.php
#	src/Tasks/Backup/DbDumperFactory.php
#	src/Tasks/Backup/Zip.php
#	src/Tasks/Cleanup/CleanupJob.php
#	src/Tasks/Cleanup/Strategies/DefaultStrategy.php
#	src/Tasks/Monitor/HealthCheck.php
#	src/Tasks/Monitor/HealthChecks/IsReachable.php
#	src/Tasks/Monitor/HealthChecks/MaximumAgeInDays.php
#	src/Tasks/Monitor/HealthChecks/MaximumStorageInMegabytes.php

+++ src/BackupDestination/BackupDestination.php
@@ -2,8 +2,8 @@
 
 namespace Develoopin\Backup\BackupDestination;
 
-use Exception;
 use Carbon\Carbon;
+use Exception;
 use Illuminate\Contracts\Filesystem\Factory;
 use Illuminate\Contracts\Filesystem\Filesystem;
 use Develoopin\Backup\Exceptions\InvalidBackupDestination;
@@ -139,7 +139,7 @@ class BackupDestination
         }
     }
 
-    public function usedStorage(): int
+    public function usedStorage(): float
     {
         return $this->backups()->size();
     }
